Move heart of the code back into Addon folder
