+++ package.json
@@ -30,7 +30,6 @@
     "ember-cli-htmlbars-inline-precompile": "^0.3.3",
     "ember-cli-inject-live-reload": "^1.4.1",
     "ember-cli-jshint": "^1.0.4",
-    "ember-cli-mirage": "^0.2.4",
     "ember-cli-mocha": "0.13.1",
     "ember-cli-release": "^0.2.9",
     "ember-cli-showdown": "2.8.0",
